+++ CHANGELOG.md
@@ -10,6 +10,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * `kwarray.embed_slice` now does error checking
 * Add `kwarray.RunningStats.update_many` to update with multiple observations at once.
 
+### Fixed
+* Fix ArrayAPI.max for torch in the case where not both items are ints
+
 
 ## Version 0.6.3 - Released 2022-07-31